Added new components for language information and changed variable structure
