+++ src/app/LanguageSources.svelte
@@ -1,0 +1,3 @@
+<div>
+  <h1>Language Sources Component</h1>
+</div>
